Revert "Refactor layouts, integrate Unleash, and remove whiteboard feature"
