--- .env.example
@@ -1,2 +1,0 @@
-PUBLIC_UNLEASH_URL=http://localhost:4242/api/frontend
-PUBLIC_UNLEASH_CLIENT_KEY=default:development.unleash-insecure-frontend-token
