+++ package.json
@@ -17,7 +17,7 @@
 		"@sveltejs/vite-plugin-svelte": "^6.2.0",
 		"@tailwindcss/typography": "^0.5.18",
 		"@tailwindcss/vite": "^4.1.13",
-		"@types/node": "^24.10.1",
+		"@types/node": "^24.9.1",
 		"svelte": "^5.39.5",
 		"svelte-check": "^4.3.2",
 		"tailwindcss": "^4.1.13",
@@ -27,7 +27,6 @@
 	"dependencies": {
 		"@sveltejs/adapter-node": "^5.4.0",
 		"@unleash/proxy-client-svelte": "^0.4.3",
-		"js-yaml": "^4.1.0",
-		"unleash-client": "^6.8.0"
+		"js-yaml": "^4.1.0"
 	}
 }
