+++ src/app.d.ts
@@ -3,11 +3,7 @@
 declare global {
 	namespace App {
 		// interface Error {}
-		interface Locals {
-			flags: {
-				isEnabled: (name: string) => boolean;
-			};
-		}
+		// interface Locals {}
 		// interface PageData {}
 		// interface PageState {}
 		// interface Platform {}
