--- src/routes/app/+page.svelte
@@ -1,10 +1,0 @@
-<script>
-    import { goto } from '$app/navigation';
-    import { onMount } from 'svelte';
-
-    onMount(() => {
-        goto('/app/projects');
-    });
-</script>
-
-<p>Redirecting...</p>
