--- src/routes/app/settings/+page.svelte
@@ -1,2 +1,0 @@
-<h1>Settings</h1>
-<p>Manage your preferences here.</p>
