+++ src/routes/whiteboard/+page.server.ts
@@ -1,6 +1,5 @@
 import type { PageServerLoad } from './$types';
 import { BACKEND_DOMAIN_NAME, BEARER_TOKEN_BACKEND } from '$env/static/private';
-import { unhashUserID } from '$lib/server/auth';
 
 interface Project {
     created_at: string;
@@ -11,14 +10,8 @@ interface Project {
     submission_week: number;
 }
 
-export const load: PageServerLoad = async ({ cookies }) => {
-    const hashedUserID = cookies.get('userID');
-
-    if (!hashedUserID) {
-        return { projects: [] as Project[] };
-    }
-
-    const userID = unhashUserID(hashedUserID);
+export const load: PageServerLoad = async ({ parent }) => {
+    const { userID } = await parent();
 
     try {
         // Fetch user's projects
