+++ src/routes/whiteboard/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-    import type { PageData } from './$types';
     import { onMount } from 'svelte';
+    import type { PageData } from './$types';
     
     let { data }: { data: PageData } = $props();
     
@@ -8,7 +8,26 @@
     let innerHeight = $state(0);
     let mediumSize = $derived(innerWidth < 1850 || innerHeight < 1100);
 
+    let hackatimeAcknowledged = $state(false);
+
+    function acknowledgeHackatime() {
+        hackatimeAcknowledged = true;
+        document.cookie = "hackatimeAcknowledged=true; path=/; max-age=" + (30 * 24 * 60 * 60);
+    }
+
+    function getCookie(name: string) {
+        const value = `; ${document.cookie}`;
+        const parts = value.split(`; ${name}=`);
+        if (parts.length === 2) return parts.pop()?.split(';').shift();
+        return null;
+    }
+
     onMount(() => {
+        const acknowledged = getCookie('hackatimeAcknowledged');
+        if (acknowledged === 'true') {
+            hackatimeAcknowledged = true;
+        }
+
         const updateSize = () => {
             innerWidth = window.innerWidth;
             innerHeight = window.innerHeight;
@@ -22,18 +41,18 @@
         };
     });
 </script>
-
-<div class="container mx-auto pt-10">
+{#if hackatimeAcknowledged}
+<div class="container mx-auto pt-30">
     <div class="mt-3">
         <h2 class="{mediumSize ? 'text-3xl' : 'text-5xl'} font-bold text-neutral-700 mb-1">Your Recent Projects</h2>
         
         {#if data.projects.length === 0}
-            <p class="{mediumSize ? 'text-xl' : 'text-3xl'} text-neutral-600">No projects found.</p>
+            <p class="{mediumSize ? 'text-xl' : 'text-3xl'} text-neutral-600">No projects</p>
         {:else}
             <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
                 {#each data.projects.sort((a, b) => new Date(b.created_at).getTime() - new Date(a.created_at).getTime()).slice(0, 3) as project}
                     <div class="block {mediumSize ? 'max-w-sm' : ''}">
-                        <div class="rounded-lg shadow-lg overflow-hidden hover:scale-105 relative bg-white">
+                        <div class="rounded-lg shadow-lg overflow-hidden hover:scale-105 relative">
                             <div class="{mediumSize ? 'p-2 pb-10' : 'p-4 pb-12'}">
                                 <a href={project.live_url} target="_blank" rel="noopener noreferrer">
                                     <h3 class="{mediumSize ? 'text-sm' : 'text-xl'} font-bold text-neutral-800 mb-2">{project.project_name}</h3>
@@ -55,3 +74,14 @@
         {/if}
     </div>
 </div>
+{:else}
+<h1 class="text-center mt-40 text-6xl font-bold">
+    Have you installed Hackatime?
+</h1>
+<div class="text-center"><button onclick={() => window.location.href ="https://hackatime.hackclub.com"} class="bg-red-700/50 px-10 py-2 text-2xl rounded-xl hover:scale-105 hover:bg-red-700/30 hover:cursor-pointer mt-10 mb-10 hover:underline"><h2>If you haven't, install it now</h2></button>
+<h2 class="text-2xl font-bold mb-10">If you have, acknowledge this:</h2>
+<p class="text-2xl font-pangolin px-20">Don’t cheat the time tracking system. No bots, no fake key presses, no UI manipulation. If you do, you’ll be banned from Hackatime and other participating YSWS / events / programs</p>
+<button onclick={acknowledgeHackatime} class="bg-green-700/50 px-10 py-2 text-2xl rounded-xl hover:scale-105 hover:bg-green-700/30 hover:cursor-pointer mt-10 mb-5 hover:underline"><p>I agree</p></button>
+<img src="/please.gif" alt="Pwease" class="w-48 mx-auto mt-10"/>
+</div>
+{/if}
